Log query warnings before setting the final result
Motivation:

Currently the callback result is set _before_ the query warnings are
logged. As a consequence, client code dependent on the callback's
completion may get executed before the warnings are logged.

Modification:

Set the callback result after the query warnings are logged.

Result:

Client code is guaranteed to be executed after the query warnings are
logged.