JAVA-2254: Add support for computed properties (#1242)
Motivation:

It is useful to map values that are the result of some Cassandra-side
computation, typically function calls, to properties.

Modifications:

Introduce a `@Computed` annotation that can be applied to a field or
getter that allows specifying a formula to instruct Cassandra to
compute and return a value back that can be mapped back to this
property.

Add EntityDefinition.getComputedValues() for retrieving computed values
and added getAllValues(), which is similar to getAllColumns() but also
returns computed values.

Add PropertyDefinition.getSelector() to differentiate between
cql selector and cql name for alias definitions.

Update EntityHelperSelectStartMethodGenerator to include computed values
in generated select query.

Result:

`@Computed` annotation may be used to specify properties that are based
on result of some function call.  This is only usable by
`@Select`-annotated dao methods.