JAVA-2111: Add HierarchyScanStrategy to configure inheritance behavior (#1250)
Motivation:

The ability to define properties and annotations of Entity-annotated
classes in parent classes or interfaces will make it easier for
implementors to reuse code and enforce behavior across multiple
entities.

In addition, the same applies for Dao-annotated interfaces.  For
example, being able to express a NullSavingStrategy on a shared parent
interface would be desirable for those who wish to share that
configuration among multiple Daos.

Modifications:

Added `@HierarchyScanStrategy` that can be defined anywhere in a Dao or
Entity type hierarchy.  This strategy dictates how to treat annotations
and properties defined in parent classes and interfaces.

By default, all parent interfaces and classes are considered.

Result:

The mapper can now make use of properties and annotations defined in
parent classes and interfaces of Dao and Entity-annotated types.