JAVA-2288: Improve error reporting when DAO is missing keyspace
Motivation:

If no keyspace can be discerned from the session's keyspace, the
entity's default keyspace, or the dao's parameters, the mapper should
preemptively throw a RuntimeException with a helpful message to guide
the user to provide a keyspace.

Modifications:

Added throwIfKeyspaceMissing call in Dao initAsync method to check if
keyspace is set.

Result:

User is now presented with a helpful exception message when keyspace is
not configured.