JAVA-2297: Set socket options from NettyOptions#afterBootstrapInitialized
Motivation:

Setting socket options on the bootstrap is currently hardcoded in
ChannelFactory. This can cause warnings if the transport is overridden
with an implementation that does not support some of these options (ex:
TCP_NODELAY for Unix domain sockets).

Modifications:

Move the code to NettyOptions#afterBootstrapInitialized.

Result:

Custom NettyOptions implementations can decide which socket options to
set. Note that they will now have to copy over the code from
DefaultNettyOptions to get the default behavior.