JAVA-2315: Improve extensibility of session builder
Motivation:

Every time we add a new programmatic argument to the session builder,
there is a breaking change on SessionBuilder#buildContext.

This extends to DefaultDriverContext's constructor (even though
DefaultDriverContext is an internal class, it would be wise to be
conservative about its API, because plugging something through a custom
context subclass is one of the most basic "advanced" extensions).

Modifications:

Create a new class ProgrammaticArguments to wrap the arguments to
buildContext. Also pass it to the context constructor.

Preserve backward compatibility with the previous signatures.

Result:

We can now add new arguments to ProgrammaticArguments, which is not a
breaking change.