JAVA-2308: Add customWhereClause to Delete
Motivation:

CQL DELETE allows advanced capabilities for deleting a range of data
within a partition.  For example given a primary key of ((int k), int c),
one could perform the following:

DELETE FROM tbl WHERE k = 0 and c > 1 and c < 4

To support this capability in the mapper, customWhereClause should be
added to Delete in a similar way it is handled in the Select annotation.

Modifications:

Add customWhereClause to Delete annotation.

Update DaoDeleteMethodGenerator to handle customWhereClause.

Add EntityHelperDeleteStartMethodGenerator which adds a deleteStart
method, which is reused in
EntityHelperDeleteByPrimaryKeyPartsMethodGenerator and also used in the
code generation in DaoDeleteMethodGenerator

Result:

customWhereClause added to Delete.