JAVA-2302: Better target mapper errors and warnings for inherited methods
Motivation:

If
- the mapper processes a class or interface that inherits from another
  type
- that parent type is not part of the current compilation round (e.g.
  it's in a JAR dependency of the current project)
- that parent type is not annotated correctly

Then the errors or warnings issued by the mapper processor do not
provide enough context, because they can't be linked to a source file
and line number.

Modifications:

Detect that situation in DecoratedMessager, and issue the messages on
the child type instead.

Result:

The errors or warnings will appear on a source file, with a contextual
prefix like "[getId() inherited from BaseEntity]".