Don't reference current version in POM snippets
The latest version is now displayed in the root README thanks to the
maven badge. We don't need to repeat it in every POM snippet. Show a
Maven variable instead: it should be pretty clear that it's a
placeholder.

That way we don't need to search and replace through the whole manual
before each release.