Generify AdminRequestHandler
Motivation:

We now have a need to read the response for message types other than
ROWS. AdminResult is not well-suited for this; we could introduce
subclasses, but that would require awkward casting.

Modifications:

Generify AdminRequestHandler over the element type of the returned
future. That type is provided at construction time (this is abstracted
behind the static factory methods).

Result:

Clients use the factory method matching their expected response type;
they get a future that contains the right object directly.