Added Failover Policy
Motivation:

Failover is a critical aspect of the database layer that is missing in the current
implementation of the driver. While a cloud model does lend itself to marking down
a full site, other deployment models do not. The goal of the failover policy is to
build a simple tie in to allow for failvoer decisions to be made based on caught
exceptions as well as statement attributes.

The included DefaultFailoverPolicy allows failover to a second policy which could
then be a new dc, consistency, or many other variations.

Modifications:

Added new failover parameters, FailoverPolicy interface, and DefaultFailoverPolicy
objects. This was integrated to the CqlRequestHandler#handle method by adding a
new exceptionally stage which evaluates if a failover should occur then triggers
a new CqlRequestHandler with the modified statement.

Result:

New FailoverPolicy is available to failover statements in the event of a failure.