JAVA-2455: Improve logging of schema refresh errors
Motivation:

While debugging JAVA-2454, we noticed that the error wasn't surfaced in
the logs as expected. MetadataManager.startSchemaRequest has a listener
for that purpose, but it's on the wrong future; parsing errors will
surface on currentSchemaRefresh.

Modifications:

Don't try to log errors in MetadataManager. Instead, handle it on all
call sites of MetadataManager.refreshSchema, *if* the future is not
returned to the client.

Result:

If a user-initiated refresh fails, they will get back a failed future.
If an internal refresh fails, it will be logged.