JAVA-2464: Fix initial schema refresh when reconnect-on-init is enabled
Motivation:

When reconnect-on-init is enabled and the driver does not connect
immediately, the schema metadata is empty.

This is because the initial schema refresh is chained on
ControlConnection.firstConnectionAttemptFuture. The goal was to allow
the session to initialize without waiting for the control connection (if
there is a custom topology monitor that uses another mechanism). But in
the scenario above, firstConnectionAttemptFuture is failed.

Modifications:

Chain the schema refresh to ControlConnection.initFuture (complete,
successful connection).
Remove firstConnectionAttemptFuture.

Result:

The schema is present at startup.

Note that if a custom topology monitor is in place, and it is the
metadata manager that initializes the control connection, the first
schema refresh (and therefore the whole session initialization) will
have to wait for the control connection to successfully connect. All
things considered, this is pretty logical.