JAVA-2546: Abort schema refresh if a query fails (#1365)
* Always warn on schema query errors

The `warnIfMissing` flag was introduced to handle inconsistencies in
some DSE versions, but since JAVA-2471 we check that beforehand and
select the appropriate `SchemaQueries` implementation.

* JAVA-2546: Abort schema refresh if a query fails

Motivation:

The current strategy is "best effort": ignore query errors and keep
parsing as much as we can.

However this has undesirable consequences: we observed a case where the
query to `system_schema.keyspaces` times out. Parsing proceeds with an
empty keyspace list, `SchemaRefresh` thinks all keyspaces have been
removed and notifies schema listeners.

Modifications:

If a query fails, propagate the error to the schema rows future. Ignore
any remaining queries.

Result:

The refresh is aborted and the previous version of the schema metadata
stays in place.
This is either logged, or surfaced directly if the refresh was triggered
manually (by mechanisms that were already in place before this commit).