JAVA-2473: Don't reconnect control connection if protocol is downgraded
Motivation:

In theory, the control connection is an implementation detail of the
topology monitor and metadata manager. If those components are
overridden with custom versions that use another way to get the data,
the rest of the driver should be able to function without initializing
the control connection at all.

This is currently not the case: DefaultSession references the control
connection, in order to reconnect if the protocol version was
downgraded after the initial connection (JAVA-1295).

Modifications:

Do not reconnect the control connection in that case; it's not really
needed because none of the control queries use any protocol-dependent
feature.

This also allows us to remove MetadataManager.firstSchemaRefreshFuture()
and slightly simplify session initialization.

Result:

DefaultSession does not reference ControlConnection.

If the protocol version is downgraded, the control connection keeps
using the initial version. It might switch to the "correct" version
later if it reconnects to another node.