JAVA-2442: Don't check for schema agreement twice when completing a DDL query
Motivation:

When we process a SCHEMA_CHANGE response to a DDL query, we check for
schema agreement, and then trigger a schema refresh. But the refresh
itself also checks for agreement, so we end up doing it twice.

Modifications:

Only check for agreement during the refresh. Introduce a new
RefreshSchemaResult type, so that we can surface the outcome of the
agreement in addition to the new metadata.
Remove the agreement check in CqlRequestHandler, use the value surfaced
by the refresh instead.

Result:

Schema agreement runs only once.