JAVA-2637: Bump Netty to 4.1.45
This required adjustements in a few unit tests, where
`waitForPendingAdminTasks` methods relied on the fact that a task
scheduled on the event loop with a short delay would be executed after
any current immediate task.

This doesn't work anymore, possibly after the event loop refactoring in
Netty 4.1.44. It was replaced with a combination of Awaitility and
timeouts in `Mockito.verify` calls.