Explicitly include org.json:json as direct dependency
org.json:json is a transitive dependency of ESRI.
ESRI pulls in the version 20090211, which not only
is very old, but also has the inconvenience of
not being an OSGi bundle, and not having attached
sources. Using a more recent version (20190722)
allows us to generate better javadocs for the
shaded jar, and to run OSGi tests that exercise
geotypes.