Fix impl. of getFormattedContents for TupleValue and UdtValue
Without this fix, the codec registry might return a registered
custom codec for the CQL type instead of generating a new
codec on the fly for tuple <-> TupleValue or udt <-> UdtValue.

If the registered custom codec does not handle TupleValue or
UdtValue, the call to TypeCodec.format() will throw a
ClassCastException.