Update error-prone to 2.3.4
This commit upgrades error-prone to 2.3.4 and updates all
the sites where new problems were reported.

This is mainly motivated for compatibility with IntelliJ's
error-prone plugin, that was recently upgraded to 2.3.4.

Mockito warnings were removed from error-prone core, and
thus the MockitoInternalUsage warning exception was removed
from the Maven configuration.

Instead, two new warnings are now turned OFF:

- PreferJavaTimeOverload reports dozens of false positives
in the AssertJ API.
- AnnotateFormatMethod is useless in the context of the
driver and requires additional dependencies to be satisfied.