JAVA-2837: make StringCodec strict about unicode in ascii
Somewhere between 3.x and 4.x the StringCodec implementation started
using String.getBytes(charset), which has the caveat of translating
unmappable characters into a charset-dependent replacement character,
which for ascii is '?'. In other words, if you were to put unicode data
into an ascii field, it will just insert a lot of question marks.

With this patch the driver will throw an InvalidArgumentException if
that happens.