Remove explicit dependency to javatuples
It is a transitive dependency of Tinkerpop. We do reference it directly
from the driver code (CqlCollectionPredicate), so in theory we should
redeclare it, but it's only used to deserialize graph stuff and not
exposed in our public API.

The advantage of not declaring it explicitly is that it's one less
thing to exclude manually if you're not using graph.