JAVA-2863: Reintroduce mapper processor dependency to SLF4J
Motivation:

In JAVA-2800 we removed that dependency to avoid SLF4J's "defaulting to
NOP" warning when the processor in configured via `-processorpath`.
However, the processor depends on driver core classes that sometimes
declare a logger. If SLF4J is missing, this will cause a
ClassNotFoundException while processing annotations.

Some more background:

- in theory the processor should be completely isolated from SLF4J (its
  only output is compiler messages). But sometimes reusing core classes
  is convenient, we wouldn't want to rewrite them from scratch. A NOP
  implementation that ignores the logs is what we want, it's just the
  warning that is undesirable.
- we can't declare slf4j-nop as an explicit dependency of the processor,
  because then it would collide with the application's binding if the
  processor is configured via the regular classpath instead of
  `-processorpath`.

Modifications:

Revert JAVA-2800 to reintroduce the dependency.
Add a snippet in the manual to explain how users can add slf4j-nop to
their processorpath to avoid the "defaulting to NOP" warning.

Result:

No more risk of ClassNotFoundException.
The warning is back, but it's not harmful and users have a workaround to
remove it.