Skip should_set_all_occurrences_of_variable() test
Skip should_set_all_occurrences_of_variable() test of
BoundStatementCcmIT due to scylladb/scylla#10956. It was decided that
Scylla will have different behavior compared to Cassandra when there are
multiple uses of the same named variable - only a single copy of that
value will be sent by the driver and the driver will receive a single
"slot" for that value.