Switch from Docker Registry v1 API to v2 API in CI
On September 5, 2022 Docker Hub deprecated and disabled their
Docker Registry v1 API endpoint. We used this endpoint as a simple way
to fetch a list of Scylla OSS and Scylla Enterprise tags in order to
determine the latest versions of released Scylla images to test against.

This change caused our CI to fail to test Scylla versions as it was 
unable to determine the latest release version numbers.

To solve this problem, ci/version_fetch.py script is updated to use the
Docker Registry v2 API endpoint. The new API is paginated, necessitating
performing multiple HTTP requests in fetch_docker_hub_tags() function.