Removed unneeded GraalVM delete substitutions
@Delete is not supported (native build broken) on classes
that have a field of their own type with GraalVM 22.2 and
upcoming 22.3, and that's the case for Lz4Compressor which
has a Lz4Compressor field.

In practice, those @Delete substitutions are useless because
Lz4Compressor and SnappyCompressor won't be reachable after
newInstance substitution has been applied, so this commit
removes them.