Fix Dependency was unintentionally initialized at build time error
The Dependency class is used in the various BooleanSupplier
implementations for the class specified in the @TargetClass
onlyWith annotation attribute.

Without an explicit initialization of this class at build time,
the native build is broken with a "Dependency was unintentionally
initialized at build time error" message.

This commit initializes the Dependency class at build time, which
seems to be safe in terms of compatibility as it does not require
transitively the initialization of other classes at build time.