Skip handle_id_changes_on_reprepare() for Scylla
Adds ScyllaSkip annotation to PreparedStatementIT.handle_id_changes_on_reprepare()
This is a test method that comes with the most recent merge with 4.18.0

It checks if in case of performing `USE <keyspace>` the server side will not
reprepare the statement under different ID if keyspace was already defined
in query string. It seems it works this way in select versions of Cassandra
up to 4.0.1 and every higher than that.