Add USING TIMEOUT to CassandraSchemaQueries
The driver will now use existing configurable option
`METADATA_SCHEMA_REQUEST_TIMEOUT` in a `USING TIMEOUT` clause
with each schema query (if applicable).
This will make it a server side timeout in addition to driver side.
The driver will apply the `USING` clause if the sharding info of control
connection channel is not null. This is used as a proxy check to differentiate
between Scylla and Cassandra.
The behaviour should remain unchanged for Cassandra clusters.
USING TIMEOUT is a ScyllaDB CQL extension and it is available in
versions 4.4/2022.1.0 or newer.