Update LZ4 and Netty dependencies for CVE response
The primary goal here is to address CVE-2025-12183. Netty includes a
dependency on vulnerable versions of lz4-java, so update to a fixed
version of Netty as well. On the C* server side, we opted to move to the
new community fork of lz4-java, so match that decision here
(CASSANDRA-21052).

patch by Abe Ratnofsky; reviewed by Francisco Guerrero for CASSJAVA-113